Making progress....
- Completed import successfully.
Biggest hurdle was getting the SQL syntax correct, otherwise completed the program in about 15 minutes.

- Completed search.php
Hardest part here so far was getting the SQL syntax correct - again. After spending a lot of searching and reading understanding how the $_GET variable works and how I wanted to use it for the SQL search.