All done!!!
- search.php modified
Re-wrote search so it looks in 5 columns - postal code, city, states (abbreviated and full) and counties. It also now breaks out the search parameters into individual strings and searches on each one.

- styles.css modified
Added a minimum width to the .label class.

- scripts.js modified
Added a labelAnchor to the marker so the place names are now centered under each one (unless it's a really long name).
Modified the template for the typeahead box to include the admin_name2 field which is typically state counties.